Create alan_partridge
Create alan_partridge file and added solution code for Codewars:

Given an array of terms, if any of those terms relate to Alan Partridge, return Mine's a Pint!

The number of exclamation mark (!) after the t should be determined by the number of Alan related terms you find in the given array (x).